rf: Split table rules into separate module
